Finalize new TLReader
